In tests, ignore minor regression in click 7.0
click 7.0 does not print the label of a progressbar when it's
outputting to something that is not a tty. Instead, it prints
nothing. However, this is a minor cosmetic issue that will most
likely only ever affect tests, so it should be safe to just
skip those asserts if we are on the version exhibiting this bug.