This change allows users to define custom resource files by placing a resource file of the same name in their project directory. Currently it is enabled only for boards.json and fpga.json but can easily be enabled for other resources.
This way, users can use their own custom board without having to burden the stock resource files with niche boards.

The bulk of the change here is making the project dir known in resources.py when the resoruces are loaded.

If a custom resource file is used, an info message is printed to facilitate in supporting users.