Completed the transition of the SConstruct file to pylint. Remaining errors where fixed and/or supressed and the SConstruct files where added to the 'make lint' target.
E0611 and E1101 are due to the magic that scons does when running SConstruct, for example, improting methods and variables implicitly such that pylint is not aware of it. A partial workaround is to refer to them through the env object, for example 'env.GLOB()' instead of 'GLOB()' but it's not clear yet what is the best approach.